extract all db adapters in shared getDatabaseAdapter.js instead of defining them twice
